Fixed warning caused by last commit
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk/pwsafe/pwsafe@3517 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/ui/wxWidgets/PWSgrid.cpp
@@ -224,7 +224,7 @@ void PWSGrid::Remove(const uuid_array_t &uuid)
     DeleteRows(row);
 
     //move all the rows up by 1
-    for (int newRow = row+1; newRow < m_row_map.size(); ++newRow) {
+    for (size_t newRow = row+1; newRow < m_row_map.size(); ++newRow) {
       m_row_map[newRow-1] = m_row_map[newRow];
     }
